Add: Support for (near) bindless textures

+++ LitterboxEngine/Graphics/GHAL/Buffer.cs
@@ -8,7 +8,7 @@ public abstract class Buffer : IDisposable
 }
 
 public record BufferDescription(
-    uint Size,
+    ulong Size,
     BufferUsage Usage);
 
 public enum BufferUsage
